Move objects to avoid circular imports later
In order to avoid circular imports between interfaces and credentials,
move CredentialType and IdentityHandle to a new module
